Fix Autofit Method Infinite Loop
This commit fixes #26, where the autofit method while loop iterates
indefinitely due to a logic error. This was caused by the Font object
(self.font) not being updated to reflect the new font size
(self.font_size). This meant that the text sprite wasn't actually being
changed, so the width and height were constant, causing the main if
statement to continue decreasing the font size.

This issue was resolved by creating a method (update_font) that is
called each time the update method is called.