Add Player Respawn When Off Screen
Added check to kill player by deducting 200 health (invoking respawn if
possible) when the player sprite is 200 pixels below the bottom of the
screen. This prevents premature health deductions if the player sprite
moves past the left/right/top sides of the screen temporarily to reach a
location.

Updated arguments on self.player.hit to comply with new parameters.

Fixes #34.