Resolve #40 - Button Text Not Opaque
Button text is now fully opaque, as expected.

The issue was caused by the main surface (self.image) having the alpha
set, anything blitted onto it will have the alpha applied on top of it.
A "blanket" alpha is applied.

To resolve this issue:
An additional surface had to be created for the background and then
both the background and text are blitted onto the main surface
(self.image). This allows the alphas of the text and background to
coincide.

Resolves #40.