Use Item Call Dictionary
Removed internal item tuple in RootMenu class and replaced with
item_calls dictionary in rootmenu method of Program class. Each value is
the method it should call.

This allows the program to process method
calls that can't otherwise be passed to the RootMenu class.

Temporary print statements for each method have been added to help debug
the program.