Add Cursor Menu Selection
This commit adds cursor left click detection to select menu items. When
the cursor is on a button and a left click occurs, the associated method
will be called, just like selecting with the keyboard.

An additional function has been added in _functions, return_button has
been added to facilitate this, the function simply returns the button in
a button group with the given button text.