Fix LevelMain Sprite Groups
This commit modifies the sprite groups in the LevelMain class so that
the sprites attribute is used as the main sprite group, referencing all
sprites. This sprite group is used to draw all sprites to the screen in
the level_main method in the Program instance created in the main
module, the sprites sprite group shouldn't be its update function (as it
previously did) as some sprites would break if this occurred.

The temporary allsprites sprite group has been removed as it was used as
a temporary workaround for this issue.

Enemies need a projectiles sprite group for each instance otherwise
move_projectiles would throw an error as it didn't have a projectiles
sprite group. The projectiles sprite group has been added through class
inheritance of Entity as the Player class also utilises the attribute.

Resolves #53.