Add Event Handling System
This commit implements a way to include a handle_events method even
though we don't know the specifics (e.g. keybind D to move right). Each
child class that inherits from Screen will create additonal methods
"event handlers" and append them to the event_handlers attribute which
means it will get called within the handle_events method.

The screen_index method has been removed as it didn't serve enough of a
purpose to be its own method. To implement the method's functionality on
a calling line, you only need to replace "_" with "s." therefore this
method is not necessary.

The update method has been implemented now that handle_events is a
method in the class, just like originally implemented in the root menu.

The button selected state is being considered in implementation. It has
been added to the process_next_screen method for now.

A pylint disable comment has been added to ignore redundant pylint
messages.