Rename next_screen Attribute
The next_screen attribute has been renamed to selected. This change has
been to improve readability and interpretation. next_screen has at times
made it confusing to refer to.

The process_next_screen method name has been untouched as this is the
method that directly interfaces to the wider scope and returns what
would be the next screen to display.

next_screen references in the Program
class in the main module remain as it makes more sense as at that level,
the screens are actually being switched between.