Use String for next_screen Referencing
Instead of referencing the next screen to display using an index of
the screens attribute using a next_screen_index attribute, the next
screen is explicitly specified using the next_screen attribute.

The previous next_screen method has been renamed to process_next_screen.

Indexes are no longer used as you cannot immediately determine what each
index represents and adding new screens in between will cause the
indexes to change for other screens.

The next screen is stored in the next_screen attribute explicitly, e.g.
"resume", or None if there's no next screen yet.

The next_screen attribute has a property decorator applied to check if a
valid screen is set to next_screen.

A tuple of screens available is passed to the class instead of being
defined in the constructor directly.

This change is due to happen to all applicable classes soon.