Inherit Screen into LevelPause Class
Adoption of the Screen class via inheritance has been added to the
LevelPause class. The previous LevelPause class has been renamed to
LevelPauseOld.

An issue arose when trying to change the state of the selected screen.
Selected was set to None when it should have been set to the first valid
screen in the screens tuple attribute.
This has been resolved in the class constructor where the selected
attribute is defined.

An issue arose when trying to match buttons using the methods prefixed
set_selected_ as the Button instance text attribute were in uppercase
whilst the selected attribute in LevelPause was in lowercase.

A temporary fix has been made, using functions to change the case.

A duplicate attribute definition has been removed from LevelPauseOld.