Add Pause Screen ESC Keybind
This commit adds a ESC keybind to resume the level.

An error occurred when resuming with ESC and also moving the mouse over
a different button. The selected attribute would change to "resume" but
then be overwritten by another screen name because of the mouse.

To mitigate this issue, a check has been added to deny changes to the
selected attribute when the confirmed attribute is True.

Resolves #78.