Add Item Confirm on Button Up Left Click
This commit makes it so that item confirmations in a menu using the
cursor occur on a left click button up event instead of a down event.

This conforms with how the vast majority of buttons work in the real
world. If you hold down left click, no actions are run until you let go.
You can cancel that button click by moving the cursor off the button and
letting go of the button.

Resolves #76.