Add Root Menu Option to Pause Screen
This commit adds an option to return to the root menu from the pause
screen. It also calls the return function when a level complete/fail
screen has finished running.

The button text now replaces underscores with spaces to allow for snake
case screen naming whilst making it user friendly.

Resolves #74.