Add Variable Projectile Damage
This commit allows projectiles to inflict different magnitudes of damage
to entities that it hits.
The damage attribute is used to specify the damage it inflicts.

The Enemy class's hit method now takes amount as a parameter to allow
for variable projectile damage.

The moveprojectiles method in the LevelMain class has been modified to
implement this feature and variable names and documentation has been
updated to make it easier to understand.