Add Global Saving
This commit adds a method save_global which uses the server function,
server_addentry to request to save a score and subsequently saves to the
local game.

The initial save method has been renamed to save_local.

A new method, update_alert has been created to update the alert text
sprite more easily and to reduce code repetition.

The save_score method in the Program class in the main module has been
updated to be able to call the save_global and save_local methods.

Resolves #82.