Fix Enemy Alternating Bottom Collision Detection
This commit fixes the alternating detection of bottom enemy collision
with a platform by replacing the airduration attribute which incremented
each frame the sprite wasn't on a platform with an attribute, onplatform
, which is a bool that stores if the enemy is on a platform or not, True
if on platform, False if not on platform or not sure.

jumpmomentum is only incremented if the onplatform attribute is False
instead of being incremented every game iteration. This is the main
cause of the problem.

This change has not been implemented for the player as of yet in order
to test this new change.

A small change in the move function in _level_main where an if statement
has been changed to an elif statement for consistency and efficiency.

Resolves #98.