Fix Settings Update Crash
This commit fixes an issues where the game would crash if you were in a
level and config.json was updated (in another instance).

This was caused by the constant reading on the config file in a level by
each entity, when it was updated, the entity read an incomplete file
while it was being saved, raising an exception.

To fix this, the sfx setting is only checked when starting, resuming
or restarting a level instead of checking every frame.

Resolves #112.